BE/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.1
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cohere==5.3.3
coloredlogs==15.0.1
dataclasses-json==0.6.5
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
Django==4.2.11
django-admin-extra-buttons==1.5.8
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-framework==0.1.0
django-tastypie==0.14.7
django-vectordb==0.1.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
executing==2.0.1
fastapi==0.110.3
fastavro==1.9.4
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gevent==24.2.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
hnswlib==0.8.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
install==1.3.5
ipdb==0.13.13
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jq==1.7.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.1.46
langchain-openai==0.0.6
langsmith==0.1.52
lockfile==0.12.2
Logbook==1.7.0.post0
maildump==1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mimeparse==1.6.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.4.post2
